throw exceptions on ideas not found

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { IdeaEntity } from './idea.entity';
 import { Repository } from 'typeorm';
@@ -22,20 +22,30 @@ export class IdeaService {
     }
 
     async read (id: string) {
-        return await this.ideaRepository.findOne({ where: { id }});
+        const idea = await this.ideaRepository.findOne({ where: { id }});
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
+        return idea;
     }
 
     // using Partial on IdeaDTO as the user may update 
     // either the idea or description or both.
     async update (id: string, data: Partial<IdeaDTO>) {
+        const idea = await this.ideaRepository.findOne({ where: {id} });
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.update({ id }, data);
-        return await this.ideaRepository.findOne({ id });
+        return idea;
     }
 
     async destroy (id: string) {
+        const idea = await this.ideaRepository.findOne({ where: {id} });
+        if (!idea) {
+            throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+        }
         await this.ideaRepository.delete({ id });
-        return {
-            deleted: true
-        };
+        return idea;
     }
 }
